[ET-VK][EZ] Split `Resource.h` into multiple files within `resource/`
I'm trying to learn what's going on within this `api/` folder.

`Resource.*` files can be split to be less intimidating to readers. I'm also thinking we can flatten the hierarchy more in future changes.

Differential Revision: [D57126893](https://our.internmc.facebook.com/intern/diff/D57126893/)

[ghstack-poisoned]
